rag-service/requirements.txt
fastapi
uvicorn
pinecone-client
google-generativeai
python-dotenv
sqlalchemy
psycopg2-binary
